requirements.txt
flask
yfinance
pandas
numpy
tensorflow-cpu
scikit-learn
python-telegram-bot
requests
beautifulsoup4